NEEDS-WORK: remove now-dead About navigation code
We probably want to reinstate this. This code needs to be redone to
reflect how things are done in the Jekyll world, and we do want to keep
previously-working links to keep working (for example,
https://git-scm.com/about/info-assurance#staging-area would redirect to
https://git-scm.com/about/staging-area).

In any case, the removal was incomplete, as the `ready` function still
tries to call `AboutContent.init();` and will now fail.